pyproject.toml
[project]
name = "jam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce==2.5.2",
    "pytmx==3.32",
    "ujson==5.10.0",
]

[tool.ruff]
lint.ignore = ["F401", "F403"]